refactor: renamed global_vars to job_container
