+++ game_share_bot/scheduler/scheduler.py
@@ -1,7 +1,7 @@
 from apscheduler.schedulers.asyncio import AsyncIOScheduler
 from sqlalchemy.ext.asyncio import async_sessionmaker
 
-from game_share_bot.scheduler.global_vars import job_container
+from game_share_bot.scheduler.job_container import job_container
 from game_share_bot.scheduler.job_list import JOBS
